+++ example/src/core/config.dart
@@ -2,6 +2,6 @@ abstract interface class Config {
   const Config();
 }
 
-final class Config$ implements Config {
-  const Config$();
+final class MyConfig implements Config {
+  const MyConfig();
 }